refactoring the project
