+++ lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:bytebank/screens/transferencia/lista.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(BytebankApp());
@@ -8,149 +9,15 @@ class BytebankApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: Scaffold(
-        body: ListaTransferencias(),
-      ),
-    );
-  }
-}
-
-class FormTransferencia extends StatelessWidget {
-  final TextEditingController _controladorCampoNumeroConta = TextEditingController();
-  final TextEditingController _controladorCampoValor = TextEditingController();
-
-  FormTransferencia({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Criando Transferência'),
-      ),
-      body: Column(
-        children: [
-          Editor(
-            controller: _controladorCampoNumeroConta,
-            rotulo: 'Número da conta',
-            dica: '0000',
-          ),
-          Editor(
-            controller: _controladorCampoValor,
-            rotulo: 'Valor',
-            dica: '0.00',
-            icone: Icons.monetization_on,
-          ),
-          ElevatedButton(
-            onPressed: () => _criaTransferencia(context),
-            child: Text('Confirmar'),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _criaTransferencia(BuildContext context) {
-    final String numeroConta = _controladorCampoNumeroConta.text;
-    final String valor = _controladorCampoValor.text;
-    final transferenciaCriada = Transferencia(valor, numeroConta);
-    debugPrint('Criando Tranferencia');
-    ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text('$transferenciaCriada')));
-    Navigator.pop(context, transferenciaCriada);
-  }
-}
-
-class Editor extends StatelessWidget {
-  final TextEditingController? controller;
-  final String? rotulo;
-  final String? dica;
-  final IconData? icone;
-
-  Editor({
-    this.controller,
-    this.rotulo,
-    this.dica,
-    this.icone,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(16.0),
-      child: TextField(
-        controller: controller,
-        style: TextStyle(fontSize: 24.0),
-        decoration: InputDecoration(
-          icon: icone != null ? Icon(icone) : null,
-          labelText: rotulo,
-          hintText: dica,
+      theme: ThemeData(
+        primaryColor: Colors.green[900],
+        accentColor: Colors.blueAccent[700],
+        buttonTheme: ButtonThemeData(
+          buttonColor: Colors.blueAccent[700],
+          textTheme: ButtonTextTheme.primary,
         ),
-        keyboardType: TextInputType.number,
       ),
+      home: ListaTransferencias(),
     );
   }
-}
-
-class ListaTransferencias extends StatefulWidget {
-  @override
-  _ListaTransferenciasState createState() => _ListaTransferenciasState();
-}
-
-class _ListaTransferenciasState extends State<ListaTransferencias> {
-  final List<Transferencia?> _transferencias = List.empty(growable: true);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Transferências'),
-      ),
-      body: ListView.builder(
-        itemCount: _transferencias.length,
-        itemBuilder: (context, indice) {
-          final transferencia = _transferencias[indice];
-          return ItemTransferencia(transferencia!);
-        },
-      ),
-      floatingActionButton: FloatingActionButton(
-        child: Icon(Icons.add),
-        onPressed: () {
-          final Future<Transferencia?> future = Navigator.push(context, MaterialPageRoute(builder: (context) {
-            return FormTransferencia();
-          }));
-          future.then((transferenciaRecebida) {
-            debugPrint('chegou no then do future');
-            debugPrint('$transferenciaRecebida');
-            _transferencias.add(transferenciaRecebida);
-          });
-        },
-      ),
-    );
-  }
-}
-
-class ItemTransferencia extends StatelessWidget {
-  final Transferencia _transferencia;
-
-  ItemTransferencia(this._transferencia);
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: ListTile(
-        leading: Icon(Icons.monetization_on),
-        title: Text(_transferencia.valor.toString()),
-        subtitle: Text(_transferencia.conta.toString()),
-      ),
-    );
-  }
-}
-
-class Transferencia {
-  final String conta;
-  final String valor;
-
-  Transferencia(this.valor, this.conta);
-
-  @override
-  String toString() => 'Transferência (conta: $conta, valor: $valor)';
 }
