+++ lib/models/transferencia.dart
@@ -1,0 +1,12 @@
+class Transferencia {
+  final String conta;
+  final String valor;
+
+  Transferencia(
+    this.valor,
+    this.conta,
+  );
+
+  @override
+  String toString() => 'Transferência (conta: $conta, valor: $valor)';
+}
