+++ lib/screens/transferencia/lista.dart
@@ -1,0 +1,63 @@
+import 'package:bytebank/models/transferencia.dart';
+import 'package:flutter/material.dart';
+import 'formulario.dart';
+
+const _tituloAppBar = 'Transferências';
+
+class ListaTransferencias extends StatefulWidget {
+  @override
+  _ListaTransferenciasState createState() => _ListaTransferenciasState();
+}
+
+class _ListaTransferenciasState extends State<ListaTransferencias> {
+  final List<Transferencia?> _transferencias = [];
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(_tituloAppBar),
+      ),
+      body: ListView.builder(
+        itemCount: _transferencias.length,
+        itemBuilder: (context, indice) {
+          final transferencia = _transferencias[indice];
+          return ItemTransferencia(transferencia!);
+        },
+      ),
+      floatingActionButton: FloatingActionButton(
+        child: Icon(Icons.add),
+        onPressed: () {
+          Navigator.push(context, MaterialPageRoute(builder: (context) {
+            return FormTransferencia();
+          })).then(
+            (transferenciaRecebida) => _atualiza(transferenciaRecebida),
+          );
+        },
+      ),
+    );
+  }
+
+  void _atualiza(Transferencia? transferenciaRecebida) {
+    if (transferenciaRecebida != null) {
+      setState(() => _transferencias.add(transferenciaRecebida));
+    }
+  }
+}
+
+class ItemTransferencia extends StatelessWidget {
+  final Transferencia _transferencia;
+
+  ItemTransferencia(this._transferencia);
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      child: ListTile(
+        leading: Icon(Icons.monetization_on),
+        title: Text(_transferencia.valor.toString()),
+        subtitle: Text(_transferencia.conta.toString()),
+      ),
+    );
+  }
+}
